+++ test/spec/client.spec.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import Bot from '../../src/client.js'
+import { Bot } from '../../src/Bot.js'
 
 vi.mock('../../src/config.js', () => {
   const Environment = {
